Update BaldHub
fixed the "everyone for custom chars" and added a few more games to fix, plan on making it so the ws isn't toggled in any game until you first use it so it doesn't break certain games (hasn't got me banned in any games but does break your games sometimes). I'll be fixing it soon.